[IMP] sale_order_product_recommendation: slipt tests
- Improve test declaration inherability
