+++ sale_order_product_recommendation/tests/__init__.py
@@ -1,4 +1,3 @@
-
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
+from . import test_recommendation_common
 from . import test_recommendation
